Fixed IEEE 754 precision errors (using 80-bit extended precision for all arithmetic and stored computed floating-point values); All RangedGraph tests: PASS

+++ Sources/Modules/RangedGraph/RangedGraph.h
@@ -68,7 +68,7 @@ namespace ranged_graph
         friend class ::TestHelper;
 #endif
 
-        double pointsPerPixel_;
+        long double pointsPerPixel_;
         Range2D range2D_;
         Magick::Image image_;
